Move aseprite-like structs to doc-lite.h
